+++ src/model/factory.ts
@@ -1,24 +1,34 @@
 import {Map} from 'immutable';
-import {forEachOwnProperty} from 'caesium-core/lang';
+
+import {forEachOwnProperty, isDefined} from 'caesium-core/lang';
 import {ModelBase} from './base';
-import {ModelMetadata, PropertyMetadata} from './metadata';
+import {ModelMetadata, PropertyMetadata, RefPropertyMetadata} from './metadata';
 import {ModelValues} from "./values";
 
 export type ModelFactory<T extends ModelBase> = (properties: { [attr: string]: any}) => T;
 
 export interface PropertyMutation {
-    property: string;
+    /*
+     * The property (or reference) name that is being mutated
+     */
+    propName: string;
+    /**
+     * The new value of the property/reference.
+     */
     value: any;
 }
 
 export function createModelFactory<T extends ModelBase>(modelMeta: ModelMetadata): ModelFactory<T> {
     function create(args: {[attr: string]: any}) {
         var modelValues = _asMutableModelValues(_initModelValues());
-        forEachOwnProperty(args, (value, key) => modelMeta.checkHasProperty(key));
+        forEachOwnProperty(args, (value, key) => modelMeta.checkHasPropertyOrRef(key));
 
-        modelMeta.properties.forEach((prop: PropertyMetadata) => {
-            var initializer = prop.valueInitializer;
-            modelValues = initializer(modelValues, args[prop.name]);
+        modelMeta.properties.forEach((prop) => {
+            modelValues = prop.valueInitializer(modelValues, args[prop.name]);
+            if (prop.isRef) {
+                var refProp = prop as RefPropertyMetadata;
+                modelValues = refProp.refValueInitializer(modelValues, args[refProp.refName]);
+            }
         });
         return _createModel(modelMeta, _asImmutableModelValues(modelValues));
     }
@@ -40,23 +50,37 @@ export function createModelFactory<T extends ModelBase>(modelMeta: ModelMetadata
  */
 export function copyModel<T extends ModelBase>(
     model: T,
-    mutations?: PropertyMutation[]
+    mutations?: PropertyMutation[] | ModelValues
 ): T {
     var modelMeta = ModelMetadata.forInstance(model);
-    mutations = mutations || [];
+    if (!isDefined(mutations)) {
+        return _createModel(modelMeta, (model as any).__modelValues);
+    }
 
+    if (!Array.isArray(mutations)) {
+        return _createModel(modelMeta, mutations as ModelValues);
+    }
     var modelValues = _asMutableModelValues((model as any).__modelValues);
 
-    mutations.forEach((mutation) => {
-        modelMeta.checkHasProperty(mutation.property);
-        var mutator = modelMeta.propertyMutators.get(mutation.property);
-        modelValues = mutator(modelValues, mutation.value);
+    var propMutations = mutations as PropertyMutation[];
+
+    propMutations.forEach((mutation) => {
+        var property = modelMeta.properties.get(mutation.propName);
+        if (isDefined(property)) {
+            modelValues = property.valueMutator(modelValues, mutation.value);
+        } else {
+            // The property is a RefProperty.
+            var propName = modelMeta.refNameMap.get(mutation.propName);
+            var refProperty = modelMeta.properties.get(propName) as RefPropertyMetadata;
+            modelValues = refProperty.refValueMutator(modelValues, mutation.value);
+        }
     });
 
     return _createModel(modelMeta, _asImmutableModelValues(modelValues));
 }
 
-function _createModel(modelMeta: ModelMetadata, modelValues: ModelValues) {
+function _createModel(modelMeta: ModelMetadata, modelValues?: ModelValues) {
+    modelValues = modelValues || _initModelValues();
     var descriptors: PropertyDescriptorMap = {};
 
     descriptors['__metadata'] = {enumerable: false, writable: false, value: modelMeta};
@@ -66,6 +90,13 @@ function _createModel(modelMeta: ModelMetadata, modelValues: ModelValues) {
         descriptors[prop.name] = {
             enumerable: true,
             get: function() { return this.get(prop.name); }
+        };
+        if (prop.isRef) {
+            var refName = (prop as RefPropertyMetadata).refName;
+            descriptors[refName] = {
+                enumerable: true,
+                get: function() { return this.get(refName)} 
+            }
         }
     });
 
@@ -75,20 +106,23 @@ function _createModel(modelMeta: ModelMetadata, modelValues: ModelValues) {
 function _initModelValues(): ModelValues {
     return {
         initialValues: Map<string,any>(),
-        values: Map<string,any>()
+        values: Map<string,any>(),
+        resolvedRefs: Map<string,any>()
     };
 }
 
 function _asMutableModelValues(modelValues: ModelValues): ModelValues {
     return {
         initialValues: modelValues.initialValues.asMutable(),
-        values: modelValues.values.asMutable()
+        values: modelValues.values.asMutable(),
+        resolvedRefs: modelValues.resolvedRefs.asMutable()
     };
 }
 
 function _asImmutableModelValues(modelValues: ModelValues): ModelValues {
     return {
         initialValues: modelValues.initialValues.asImmutable(),
-        values: modelValues.values.asImmutable()
+        values: modelValues.values.asImmutable(),
+        resolvedRefs: modelValues.resolvedRefs.asImmutable()
     };
 }