+++ src/model/values.ts
@@ -1,11 +1,21 @@
 import {Map} from 'immutable';
 
 export interface ModelValues {
+    /**
+     * Property values, set either as:
+     * - arguments to the model factory; or
+     * - as default values provided by the property metadata
+     */
     initialValues: Map<string,any>;
+
+    /**
+     * The current values for the model properties.
+     */
     values: Map<string,any>;
+
+    /**
+     * Any resolved reference values, keyed by the reference property name
+     */
+    resolvedRefs: Map<string,any>;
 }
-
-export type ValueAccessor = (modelValues: ModelValues) => any;
-
-export type ValueMutator = (modelValues: ModelValues, value: any) => ModelValues;