+++ test/src/model/base.spec.ts
@@ -1,4 +1,4 @@
-import {Model, Property} from "../../../src/model/decorators";
+import {Model, Property, RefProperty} from "../../../src/model/decorators";
 import {ModelMetadata} from '../../../src/model/metadata';
 import {ModelBase} from '../../../src/model/base';
 import {createModelFactory} from '../../../src/model/factory';
@@ -9,7 +9,12 @@ abstract class MyModel extends ModelBase {
     @Property({codec: str})
     prop: string;
 
+    @RefProperty({refName: 'ref'})
+    refId: string;
+
+    ref: string;
 }
+
 
 
 export function modelBaseTests() {
@@ -21,7 +26,11 @@ export function modelBaseTests() {
             expect(instance.get('prop')).toBe('hello world');
         });
 
-        
+        it('should not be possible to directly set a value on the model', () => {
+            var instance = factory({prop: 'hello world'});
+            expect(() => { instance.id = 4239 }).toThrow();
+            expect(() => { instance.prop = 'hello' }).toThrow();
+        });
 
         it('should be possible to set the value of a property on a model', () => {
             var instance = factory({prop: 'hello world'});
@@ -36,6 +45,23 @@ export function modelBaseTests() {
             expect(() => instance.get('nonExistentProp')).toThrow();
             expect(() => instance.set('nonExistentProp', 'goodbye')).toThrow();
         });
+
+        it('should be possible to set the value of a reference on a model', () => {
+            var instance = factory({refId: 28});
+            instance = instance.set('ref', {id: 42}) as MyModel;
+
+            expect(instance.refId).toBe(42, 'Should set the propId when setting the ref');
+            expect(instance.ref).toEqual({id: 42}, 'Should also set the ref');
+
+            instance = instance.set('ref', null) as MyModel;
+            expect(instance.ref).toBe(null);
+            expect(instance.refId).toBe(null);
+
+            instance = instance.set('refId', 666) as MyModel;
+            expect(instance.refId).toBe(666);
+            expect(instance.ref).toBe(undefined, 'Setting the id should clear the resolved value');
+
+        })
 
     });
 }